Added docker-compose file to help users get started with openboxes locally (authored by jplaxco@usagimed.com)

+++ grails-app/domain/org/pih/warehouse/inventory/InventoryLevel.groovy
@@ -72,7 +72,7 @@ class InventoryLevel {
         preferred(nullable: true)
 	}
 
-    def statusMessage(Long currentQuantity) {
+    def statusMessage(Integer currentQuantity) {
         return InventoryUtil.getStatusMessage(status, minQuantity, reorderQuantity, maxQuantity, currentQuantity)
     }
